Nord 3.17.4 update
function server_load - have not found a way to query the api directly by hostname, but "server id" will work.  As a workaround will search the local "nordserversfile" to find the server id, and use that to query the api for the current server load.  See customization option "nordserversfile" to create the file.

https://github.com/ph202107/nordlist/issues/6